signal blacklists, SIGTerm handling

+++ main.cpp
@@ -1,8 +1,13 @@
 #include <cstdarg>
 #include <string.h>
 #include <algorithm>
+#include <future>
+#include <pthread.h>
+#include <fcntl.h>
 #include <sys/wait.h>
+#include <sys/mman.h>
 #include <unistd.h>
+#include <assert.h>
 #include <jsoncpp/json/json.h>
 #include "logging.hpp"
 
@@ -10,32 +15,49 @@ using namespace std;
 using namespace logging;
 
 
-
 struct TAppConfig
 {
     string executable;
     string name;
+    string signal_blacklist = "";
     vector<string> parameter;
     int signal;
     pid_t pid;
 };
 
+typedef struct
+{
+  bool done;
+  pthread_mutex_t mutex;
+} shared_data;
+
+static shared_data* data = NULL;
+
 vector<TAppConfig> config;
-int grace_period;
-string loglevel;
+int grace_period = 0;
+int handle_SIGInterrupt = 0;
+string sloglevel;
 pid_t ppid;
+pid_t kpid;
+pthread_mutex_t mutex;;
 
 void forkel (int);
+void shutdown ();
 void cleanup (bool killall = false);
 void displayCfg(const Json::Value &cfg_root);
 void read_config (string copt);
 void sig_handler(int signo);
+bool check_blacklist (string, int);
+void keepalive();
 string str_pf (string format, ...);
+void initialise_shared();
 void printVector();
 void testExec(int);
 
 int main(int argc, char **argv, char **env)
 {
+    initialise_shared();
+    pthread_mutex_lock(&data->mutex);
     ppid = getpid();
     logging::setLogLevel ("DEBUG");
     int c;
@@ -59,7 +81,7 @@ int main(int argc, char **argv, char **env)
     }
 
     read_config (copt);
-    logging::setLogLevel (loglevel);
+    logging::setLogLevel (sloglevel);
 
     logging::TRACE("Start ..."); //this_thread::sleep_for(chrono::milliseconds(10));
 
@@ -72,20 +94,31 @@ int main(int argc, char **argv, char **env)
         }
     }
 
+    keepalive();
+    if (ppid != getpid()) return (0);
+
     forkel(config.size());
 
     // Not for childs ... !
     if (ppid != getpid()) return (0);
 
-    wait (0);
+    wait(0);
+
+    shutdown ();
+
+    return 0;
+}
+
+void shutdown ()
+{
+    if (handle_SIGInterrupt != 0) kill(kpid, 9);
 
     cleanup ();
 
     logging::INFO(str_pf("Wait grace period: %i\n", grace_period));
     int status;
-
     chrono::steady_clock::time_point start = chrono::steady_clock::now();
-    while (waitpid(0, &status, WNOHANG) != -1)
+    while (waitpid(-1, &status, WNOHANG) != -1)
     {
         chrono::seconds duration =  std::chrono::duration_cast<std::chrono::seconds>(chrono::steady_clock::now() - start );
         if (duration.count() > grace_period) break;
@@ -94,9 +127,23 @@ int main(int argc, char **argv, char **env)
 
     cleanup (true);
 
-    return 0;
+
 }
 
+void keepalive ()
+{
+    if (handle_SIGInterrupt == 0) return;
+    if ((kpid = fork()) < 0)
+    {
+        perror("fork");
+    }
+    else
+    {
+        if (ppid != getpid()) pthread_mutex_lock(&data->mutex);
+        //if (kpid == getpid()) return;
+    }
+
+}
 void cleanup(bool killall)
 {
   for (vector<TAppConfig>::const_iterator i = config.begin(); i != config.end(); ++i)
@@ -115,14 +162,30 @@ void cleanup(bool killall)
 
 void sig_handler(int signo)
 {
-  for (std::vector<TAppConfig>::const_iterator i = config.begin(); i != config.end(); ++i)
+    if (ppid != getpid()) return;
+
+    if ((handle_SIGInterrupt != 0) && (signo ==2)) pthread_mutex_unlock(&data->mutex);
+
+
+    for (std::vector<TAppConfig>::const_iterator i = config.begin(); i != config.end(); ++i)
     {
-        pid_t pid = (*i).pid;
-        kill(pid, signo);
+        // Do not propagate blacklistet signale
+        if (!check_blacklist ((*i).signal_blacklist, signo))
+        {
+            pid_t pid = (*i).pid;
+            kill(pid, signo);
+        }
     }
 }
 
-
+bool check_blacklist (string signal_blacklist, int signal)
+{
+    string str_signal = to_string(signal);
+    if (signal < 10) str_signal = "0" + str_signal;
+    size_t found = signal_blacklist.find(str_signal);
+    if (found==std::string::npos) return false;
+    return true;
+}
 
 void read_config(string copt)
 {
@@ -132,7 +195,8 @@ void read_config(string copt)
     std::ifstream cfgfile(copt);
     reader.parse(cfgfile, root, false);
     grace_period = root["grace-period"].asInt();
-    loglevel = root["loglevel"].asString();
+    handle_SIGInterrupt = root["handle-siginterrupt"].asInt();
+    sloglevel = root["loglevel"].asString();
     Json::Value entriesArray = root["apps"];
 
     Json::Value::const_iterator i;
@@ -143,7 +207,7 @@ void read_config(string copt)
         appConfig.name = (*i)["name"].asString();
         for (Json::Value::ArrayIndex j = 0; j != (*i)["parameter"].size(); j++)
             appConfig.parameter.push_back(((*i)["parameter"][j]).asString());
-        //convertVtoC (appConfig.parameter);
+        appConfig.signal_blacklist = (*i)["signal-blacklist"].asString();
         appConfig.signal = (*i)["signal"].asInt();
         config.push_back(appConfig);
     }
@@ -179,7 +243,6 @@ void forkel(int nprocesses)
                 param_list[cntr + 1] = (char*) chars[cntr];
             param_list[cfg.parameter.size() + 1] = 0;
             logging::DEBUG(str_pf("Executable: %s   Name: %s   ParentPID: %i   PID: %i\n", executable, param_list[0], ppid, cpid));
-
             int ret = execv(executable, param_list);
             if (ret == -1) logging::ERROR(str_pf("Start of %s failed\n", param_list[0]));
             //exit(-1);
@@ -188,6 +251,8 @@ void forkel(int nprocesses)
         {
             //parent
             config[nprocesses - 1].pid = pid;
+            //Every process gets his own pid as process group
+            setpgid (pid, pid);
             forkel(nprocesses - 1);
         }
     }
@@ -209,6 +274,24 @@ string str_pf(const string fmt_str, ...) {
             break;
     }
     return std::string(formatted.get());
+}
+
+
+void initialise_shared()
+{
+    // place our shared data in shared memory
+    int prot = PROT_READ | PROT_WRITE;
+    int flags = MAP_SHARED | MAP_ANONYMOUS;
+    data = (shared_data*) mmap(NULL, sizeof(shared_data), prot, flags, -1, 0);
+    assert(data);
+
+    data->done = false;
+
+    // initialise mutex so it works properly in shared memory
+    pthread_mutexattr_t attr;
+    pthread_mutexattr_init(&attr);
+    pthread_mutexattr_setpshared(&attr, PTHREAD_PROCESS_SHARED);
+    pthread_mutex_init(&data->mutex, &attr);
 }
 
 void testExec (int num)
